+++ Client/src/components/ActionBar/ActionBar.tsx
@@ -1,4 +1,4 @@
-import { useActions } from '../hooks/useActions';
+import { useActions } from '../../hooks/useActions';
 import './ActionBar.css';
 
 interface ActionBarProps {
