+++ Client/src/components/ActionBar/index.js
@@ -1,0 +1,3 @@
+import ActionBar from './ActionBar';
+
+export default ActionBar;
