+++ Client/src/components/AddCell/index.js
@@ -1,0 +1,3 @@
+import AddCell from './AddCell';
+
+export default AddCell;
