+++ Client/src/components/CellList/CellList.tsx
@@ -1,10 +1,10 @@
 import './CellList.css';
 import { useEffect, Fragment } from 'react';
-import { useTypedSelector } from '../hooks/useTypedSelector';
-import { useActions } from '../hooks/useActions'
-import CellListItem from './CellListItem';
-import AddCell from './AddCell';
-import fetchUserCode from '../hooks/fetchUserCode'
+import { useTypedSelector } from '../../hooks/useTypedSelector';
+import { useActions } from '../../hooks/useActions'
+import CellListItem from '../CellListItem';
+import AddCell from '../AddCell';
+import fetchUserCode from '../../hooks/fetchUserCode'
 
 const CellList: React.FC = () => {
