+++ Client/src/components/CellList/index.js
@@ -1,0 +1,3 @@
+import CellList from './CellList';
+
+export default CellList;
