+++ Client/src/components/CellListItem/index.js
@@ -1,0 +1,3 @@
+import CellListItem from './CellListItem';
+
+export default CellListItem;
