+++ Client/src/components/CodeCell/index.js
@@ -1,0 +1,3 @@
+import CodeCell from './CodeCell';
+
+export default CodeCell;
