+++ Client/src/components/ElixirCodeEditor/index.js
@@ -1,0 +1,3 @@
+import ElixirCodeEditor from './ElixirCodeEditor';
+
+export default ElixirCodeEditor;
