+++ Client/src/components/JavaScriptCodeEditor/index.js
@@ -1,0 +1,3 @@
+import JavaScriptCodeEditor from './JavaScriptCodeEditor';
+
+export default JavaScriptCodeEditor;
