+++ Client/src/components/Preview/index.js
@@ -1,0 +1,3 @@
+import Preview from './Preview';
+
+export default Preview;
