+++ Client/src/components/Resizable/Resizable.tsx
