+++ Client/src/components/Resizable/index.js
@@ -1,0 +1,3 @@
+import Resizable from './Resizable';
+
+export default Resizable;
