+++ Client/src/components/RubyCodeEditor/index.js
@@ -1,0 +1,3 @@
+import RubyCodeEditor from './RubyCodeEditor';
+
+export default RubyCodeEditor;
