+++ Client/src/components/Signup/index.js
@@ -1,0 +1,3 @@
+import Signup from './Signup';
+
+export default Signup;
