+++ Client/src/components/TextEditor/index.js
@@ -1,0 +1,3 @@
+import TextEditor from './TextEditor';
+
+export default TextEditor;
