--- Client/src/components/UserCode.tsx
@@ -1,11 +1,0 @@
-import 'UserCode.css';
-
-const UserCode: React.FC = () => {
-    return (
-        <div>
-            UserCode
-        </div>
-    )
-}
-
-export default UserCode;
