+++ Client/src/components/UserCodes/UserCodes.tsx
@@ -1,0 +1,11 @@
+import './UserCodes.css';
+
+const UserCodes: React.FC = () => {
+    return (
+        <div>
+            UserCodes
+        </div>
+    )
+}
+
+export default UserCodes;
