+++ Client/src/components/UserCodes/index.js
@@ -1,0 +1,3 @@
+import UserCodes from './UserCodes';
+
+export default UserCodes;
